% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{article_listing}
\alias{article_listing}
\title{Render a listing of Radix articles}
\usage{
article_listing(collection)
}
\arguments{
\item{collection}{Collection name}
}
\description{
Render a listing of Radix articles
}
