% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_article}
\alias{create_article}
\title{Create a new article}
\usage{
create_article(file, create_dir = FALSE, edit = TRUE)
}
\arguments{
\item{file}{File name for the draft}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(defaults to \code{FALSE}).}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\description{
Create (and optionally edit) a new distill article.
}
