% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{create_theme}
\alias{create_theme}
\title{Create a Distill theme CSS file}
\usage{
create_theme(name = "theme", edit = TRUE)
}
\arguments{
\item{name}{Name of theme file (will be written as name.css)}

\item{edit}{Open an editor for the theme file}
}
\description{
Create a theme CSS file and write it to the current working directory
}
