% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_website}
\alias{create_website}
\alias{create_blog}
\title{Create a Distill website}
\usage{
create_website(dir, title, gh_pages = FALSE, edit = interactive())

create_blog(dir, title, gh_pages = FALSE, edit = interactive())
}
\arguments{
\item{dir}{Directory for website}

\item{title}{Title of website}

\item{gh_pages}{Configure the site for publishing using \href{https://pages.github.com/}{GitHub Pages}}

\item{edit}{Open site index file or welcome post in an editor.}
}
\description{
Create a basic skeleton for a Distill website or blog. Use the \code{create_website()}
function for a website and the \code{create_blog()} function for a blog.
}
\note{
The \code{dir} and \code{title} parameters are required (they will be prompted for
interactively if they are not specified).
}
\examples{
\dontrun{
library(distill)
create_website("mysite", "My Site")
}
}
