% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-package.R
\docType{package}
\name{distill-package}
\alias{distill}
\alias{distill-package}
\title{distill: 'R Markdown' Format for Scientific and Technical Writing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Scientific and technical article format for the web. 'Distill' articles feature attractive, reader-friendly typography, flexible layout options for visualizations, and full support for footnotes and citations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/distill}
  \item \url{https://pkgs.rstudio.com/distill}
  \item Report bugs at \url{https://github.com/rstudio/distill/issues}
}

}
\author{
\strong{Maintainer}: Christophe Dervieux \email{cderv@posit.co} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})

Authors:
\itemize{
  \item JJ Allaire \email{jj@posit.co} (\href{https://orcid.org/0000-0003-0174-9868}{ORCID})
  \item Rich Iannone
  \item Alison Presmanes Hill (\href{https://orcid.org/0000-0002-8082-1890}{ORCID})
  \item Yihui Xie (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item  Google LLC (https://distill.pub/guide/) [contributor, copyright holder]
  \item Nick Williams (https://wicky.nillia.ms/headroom.js/) [contributor, copyright holder]
  \item Denis Demchenko (https://github.com/lancedikson/bowser) [contributor, copyright holder]
  \item  The Polymer Authors (https://www.webcomponents.org/polyfills/) [contributor, copyright holder]
  \item Gábor Csárdi (whoami) [contributor, copyright holder]
  \item JooYoung Seo (\href{https://orcid.org/0000-0002-4064-6012}{ORCID}) [contributor]
}

}
\keyword{internal}
