% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radix_website.R
\name{radix_website}
\alias{radix_website}
\title{R Markdown site generator for Radix websites}
\usage{
radix_website(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}

\item{...}{Currently unused.}
}
\description{
R Markdown site generator for Radix websites
}
